fix function for frequency calculation round percentile rank to the nearest int

+++ core/calculator-api/src/main/java/by/modus/percentilerank/dto/PercentileRank.java
@@ -3,5 +3,5 @@ package by.modus.percentilerank.dto;
 public interface PercentileRank {
     Scorable getScorable();
 
-    double getPercentileRank();
+    int getPercentileRank();
 }
